refactor: remove DI and the pub sub client from GW service
