+++ services/gateway/package.json
@@ -22,20 +22,15 @@
   },
   "homepage": "https://github.com/cordis-lib/cordis#readme",
   "devDependencies": {
+    "@cordis/common": "workspace:^0.3.0",
     "@types/node": "^14.14.44",
     "@types/yargs": "^15.0.13",
-    "discord-api-types": "^0.12.1",
     "typescript": "^4.2.4"
   },
   "dependencies": {
     "@cordis/brokers": "workspace:^0.3.0",
-    "@cordis/common": "workspace:^0.3.0",
     "@cordis/gateway": "workspace:^0.3.0",
-    "erlpack": "github:discord/erlpack",
-    "reflect-metadata": "^0.1.13",
     "tslib": "^2.2.0",
-    "tsyringe": "^4.5.0",
-    "yargs": "^15.4.1",
-    "zlib-sync": "^0.1.7"
+    "yargs": "^15.4.1"
   }
 }
